pyproject.toml
[tool.poetry]
name = "harmony"
version = "0.7.0"
description = ""
authors = ["srntqn <serenityqn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
kubernetes = "^12.0.1"
GitPython = "^3.1.12"
PyYAML = "^5.4.1"
typer = "^0.3.2"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
requests = "^2.23.0"
kubernetes = "^12.0.1"
pycodestyle = "^2.5.0"
GitPython = "^3.1.12"
pytest = "^6.2.2"
flake8 = "^3.8.4"
build = "^0.5.1"
responses = "^0.13.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
